refactor(musefs-core): extract scan.rs inline test modules into scan/
scan.rs carried five inline `#[cfg(test)]` modules totalling ~1485 lines below
the implementation. Move each into its own file under scan/
(`scan_unit_tests`, `ogg_probe_tests`, `wav_probe_tests`, `hardening_tests`,
`bounded_probe_tests`), declared via `#[cfg(test)] mod <name>;`. scan.rs drops
2882 -> ~1406 lines.

The production code (lines <=1396) is left byte-identical, so all 16
`file:line:col` cargo-mutants anchors in .cargo/mutants.toml stay valid with no
re-anchoring (the test modules were entirely trailing). The implementation-level
concern split proposed in #445 for scan.rs is intentionally deferred: it would
shift every anchor and require manual re-anchoring via `cargo mutants --list`,
which is not justified now that facade and mp4 already capture #445's core wins.

Addresses #444 (test extraction) for scan; #445's scan impl-split deferred.
