+++ musefs-core/src/scan.rs
@@ -1395,1488 +1395,12 @@ pub fn revalidate(db: &Db, root: &Path) -> Result<RevalidateStats> {
 }
 
 #[cfg(test)]
-mod scan_unit_tests {
-    use super::*;
-    use musefs_format::PictureType;
-    use std::io::Write;
-
-    // --- ScanOptions defaults (WINDOW L16, BATCH_BYTES L12) ---
-
-    // kills the WINDOW `<<`→`>>` and BATCH_BYTES initializer mutants: the
-    // right-hand sides are decimal literals, so a mutated const/Default
-    // initializer cannot flow to both sides of the assertion.
-    #[test]
-    fn scan_options_defaults() {
-        let d = ScanOptions::default();
-        assert_eq!(d.jobs, 0, "jobs default = use available parallelism");
-        assert_eq!(d.window, 1_048_576, "window default = 1 MiB");
-        assert_eq!(d.batch_bytes, 67_108_864, "batch_bytes default = 64 MiB");
-    }
-
-    // --- read_tail_128() (lines 170-178) ---
-
-    fn write_temp(name: &str, bytes: &[u8]) -> (tempfile::TempDir, std::fs::File) {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join(name);
-        std::fs::File::create(&path)
-            .unwrap()
-            .write_all(bytes)
-            .unwrap();
-        let file = std::fs::File::open(&path).unwrap();
-        (dir, file)
-    }
-
-    // kills scan L171 `<`→`<=` (128-byte file must be Some)
-    // kills scan L172 Ok(None) constant, L178 Ok(Some) value
-    // kills scan L176 `file_len - 128`→`/` (offset 0 vs 1 shifts the bytes)
-    // kills scan L175 buf init [0;128]/[1;128] constants (exact bytes asserted)
-    #[test]
-    fn read_tail_128_exact_128_bytes() {
-        // Distinct, position-sensitive pattern: byte[i] = i (0..=127).
-        let pattern: Vec<u8> = (0u8..128).collect();
-        let (_dir, file) = write_temp("tail128.bin", &pattern);
-
-        let tail = read_tail_128(&file, 128).unwrap();
-        let expected: [u8; 128] = pattern.clone().try_into().unwrap();
-        // Exact equality kills:
-        //  - Ok(None) (would be None, not Some)
-        //  - [0;128]/[1;128] buf-init constants (would mismatch the pattern)
-        //  - `<`→`<=` (128<=128 true → returns None for a 128-byte file)
-        //  - `-`→`/` (offset 128/128==1 reads bytes[1..], shifting the pattern)
-        assert_eq!(tail, Some(expected));
-    }
-
-    // kills scan L171 `<`→`<=` boundary the other way (127 bytes → None)
-    #[test]
-    fn read_tail_128_short_file_is_none() {
-        let (_dir, file) = write_temp("tail127.bin", &[0xABu8; 127]);
-        assert_eq!(read_tail_128(&file, 127).unwrap(), None);
-    }
-
-    // --- effective_jobs() (lines 313-318) ---
-
-    // kills scan L314 effective_jobs body→1 (assuming parallelism > 1)
-    #[test]
-    fn effective_jobs_zero_uses_parallelism_and_nonzero_passes_through() {
-        let par = std::thread::available_parallelism().map_or(1, std::num::NonZero::get);
-        assert_eq!(effective_jobs(0), par);
-        assert_eq!(effective_jobs(4), 4);
-        assert_eq!(effective_jobs(1), 1);
-    }
-
-    // --- payload_weight() ---
-
-    // Sums picture + binary-tag + structural-block byte lengths (batch backpressure).
-    #[test]
-    fn payload_weight_sums_all_buffered_payloads() {
-        let pic = |n: usize| EmbeddedPicture {
-            mime: "image/png".to_string(),
-            picture_type: PictureType::new(3).unwrap(),
-            description: String::new(),
-            width: 0,
-            height: 0,
-            data: vec![0u8; n],
-        };
-        let probed = Probed {
-            format: Format::Flac,
-            audio_offset: 0,
-            audio_length: 0,
-            tags: Vec::new(),
-            pictures: vec![pic(3), pic(5)],
-            binary_tags: vec![EmbeddedBinaryTag {
-                key: "APPLICATION".into(),
-                payload: vec![0u8; 4],
-            }],
-            structural_blocks: vec![("SEEKTABLE".into(), vec![0u8; 2])],
-        };
-        // 3 + 5 (pictures) + 4 (binary) + 2 (structural) = 14.
-        assert_eq!(payload_weight(&probed), 14);
-
-        // Empty → 0, distinguishes the →1 constant (which ignores the input).
-        let empty = Probed {
-            format: Format::Flac,
-            audio_offset: 0,
-            audio_length: 0,
-            tags: Vec::new(),
-            pictures: Vec::new(),
-            binary_tags: Vec::new(),
-            structural_blocks: Vec::new(),
-        };
-        assert_eq!(payload_weight(&empty), 0);
-    }
-
-    /// Minimal-but-valid m4a that `mp4::locate_audio` accepts (one `soun` trak),
-    /// with a `udta/meta/ilst` carrying one binary `----` atom. `value` is the raw
-    /// binary `data` payload (type code 0). Not synthesis-grade (no stco), but
-    /// `probe_full` only locates audio + reads tags, never synthesizes.
-    fn mp4_with_binary_freeform(mean: &str, name: &str, value: &[u8]) -> Vec<u8> {
-        fn bx(kind: &[u8; 4], body: &[u8]) -> Vec<u8> {
-            let mut v = u32::try_from(8 + body.len())
-                .unwrap()
-                .to_be_bytes()
-                .to_vec();
-            v.extend_from_slice(kind);
-            v.extend_from_slice(body);
-            v
-        }
-        // mdia/hdlr with handler type `soun` at payload offset 8..12 (FullBox
-        // version/flags [0..4], pre_defined [4..8], handler_type [8..12]).
-        let mut hdlr_body = vec![0u8; 8];
-        hdlr_body.extend_from_slice(b"soun");
-        hdlr_body.extend_from_slice(&[0u8; 12]); // reserved(12) + empty name
-        let trak = bx(b"trak", &bx(b"mdia", &bx(b"hdlr", &hdlr_body)));
-
-        // udta/meta/ilst with one binary `----` atom.
-        let mut mean_body = 0u32.to_be_bytes().to_vec();
-        mean_body.extend_from_slice(mean.as_bytes());
-        let mut name_body = 0u32.to_be_bytes().to_vec();
-        name_body.extend_from_slice(name.as_bytes());
-        let mut data_body = 0u32.to_be_bytes().to_vec(); // type 0 = binary
-        data_body.extend_from_slice(&0u32.to_be_bytes()); // locale
-        data_body.extend_from_slice(value);
-        let mut free = bx(b"mean", &mean_body);
-        free.extend(bx(b"name", &name_body));
-        free.extend(bx(b"data", &data_body));
-        let ilst = bx(b"ilst", &bx(b"----", &free));
-        let mut meta = 0u32.to_be_bytes().to_vec();
-        meta.extend(bx(b"hdlr", &[0u8; 25]));
-        meta.extend(ilst);
-        let udta = bx(b"udta", &bx(b"meta", &meta));
-
-        let moov = bx(b"moov", &[trak, udta].concat());
-        [bx(b"ftyp", b"M4A "), moov, bx(b"mdat", b"AUDIODATA")].concat()
-    }
-
-    #[test]
-    fn probe_full_surfaces_mp4_binary_freeform() {
-        use musefs_format::mp4;
-        let bytes = mp4_with_binary_freeform("com.serato.dj", "analysis", &[0x00, 0xAB, 0xCD]);
-        let probed = probe_full(std::path::Path::new("/x.m4a"), &bytes).expect("probed");
-        assert_eq!(probed.format, Format::M4a);
-        let keys: Vec<&str> = probed.binary_tags.iter().map(|b| b.key.as_str()).collect();
-        assert!(
-            keys.contains(&"----:com.serato.dj:analysis"),
-            "binary freeform not surfaced: {keys:?}"
-        );
-        let bt = probed
-            .binary_tags
-            .iter()
-            .find(|b| b.key == "----:com.serato.dj:analysis")
-            .unwrap();
-        assert_eq!(bt.payload, vec![0x00, 0xAB, 0xCD]);
-        let scan = mp4::read_structure(&bytes).unwrap();
-        assert_eq!(probed.audio_offset, scan.mdat_payload_offset);
-    }
-
-    fn mp4_with_covr(type_code: u32, value: &[u8]) -> Vec<u8> {
-        fn bx(kind: &[u8; 4], body: &[u8]) -> Vec<u8> {
-            let mut v = u32::try_from(8 + body.len())
-                .unwrap()
-                .to_be_bytes()
-                .to_vec();
-            v.extend_from_slice(kind);
-            v.extend_from_slice(body);
-            v
-        }
-        let mut hdlr_body = vec![0u8; 8];
-        hdlr_body.extend_from_slice(b"soun");
-        hdlr_body.extend_from_slice(&[0u8; 12]);
-        let trak = bx(b"trak", &bx(b"mdia", &bx(b"hdlr", &hdlr_body)));
-
-        let mut data_body = type_code.to_be_bytes().to_vec();
-        data_body.extend_from_slice(&0u32.to_be_bytes());
-        data_body.extend_from_slice(value);
-        let ilst = bx(b"ilst", &bx(b"covr", &bx(b"data", &data_body)));
-        let mut meta = 0u32.to_be_bytes().to_vec();
-        meta.extend(bx(b"hdlr", &[0u8; 25]));
-        meta.extend(ilst);
-        let udta = bx(b"udta", &bx(b"meta", &meta));
-
-        let moov = bx(b"moov", &[trak, udta].concat());
-        [bx(b"ftyp", b"M4A "), moov, bx(b"mdat", b"AUDIODATA")].concat()
-    }
-
-    #[test]
-    fn probe_file_skips_oversized_mp4_covr() {
-        let oversized = vec![0xFFu8; MAX_ART_BYTES + 1];
-        let bytes = mp4_with_covr(13, &oversized);
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("oversized_art.m4a");
-        std::fs::write(&path, &bytes).unwrap();
-        let probed = match probe_file(&path, 0).unwrap() {
-            ProbeOutcome::Probed(p, _) => p,
-            other => panic!("expected Probed, got {other:?}"),
-        };
-        assert_eq!(probed.format, Format::M4a);
-        assert!(
-            probed.pictures.is_empty(),
-            "oversized covr must be skipped at extraction, not materialized"
-        );
-    }
-
-    #[test]
-    fn probe_file_skips_oversized_mp4_binary_freeform() {
-        // A `----` value larger than MAX_BINARY_TAG_BYTES must be skipped at
-        // extraction by the real seek-path scanner, so it is absent from Probed.
-        let oversized = vec![0xABu8; MAX_BINARY_TAG_BYTES + 1];
-        let bytes = mp4_with_binary_freeform("com.serato.dj", "analysis", &oversized);
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("oversized_bin.m4a");
-        std::fs::write(&path, &bytes).unwrap();
-        let probed = match probe_file(&path, 0).unwrap() {
-            ProbeOutcome::Probed(p, _) => p,
-            other => panic!("expected Probed, got {other:?}"),
-        };
-        assert_eq!(probed.format, Format::M4a);
-        assert!(
-            probed.binary_tags.is_empty(),
-            "oversized binary freeform must be skipped at extraction, not materialized"
-        );
-    }
-
-    #[test]
-    fn scan_options_debug_includes_progress_sink() {
-        let opts = ScanOptions {
-            progress: Some(ProgressSink::new(|_| {})),
-            ..Default::default()
-        };
-        assert!(format!("{opts:?}").contains("ProgressSink"));
-    }
-
-    #[test]
-    fn scan_emits_discovered_walked_ingested_events() {
-        use std::sync::Mutex;
-        let dir = tempfile::tempdir().unwrap();
-        for i in 0..5 {
-            let mut bytes = b"fLaC".to_vec();
-            bytes.push(0x80);
-            bytes.extend_from_slice(&[0, 0, 34]);
-            bytes.extend(std::iter::repeat_n(0u8, 34));
-            bytes.extend_from_slice(format!("AUDIO-{i}").as_bytes());
-            std::fs::write(dir.path().join(format!("t{i}.flac")), &bytes).unwrap();
-        }
-
-        let events = Arc::new(Mutex::new(Vec::<String>::new()));
-        let recorder = Arc::clone(&events);
-        let sink = ProgressSink::new(move |ev| {
-            let line = match ev {
-                ScanProgress::Discovered { found } => format!("disc:{found}"),
-                ScanProgress::Walked { total } => format!("walk:{total}"),
-                ScanProgress::Ingested { done, total, .. } => format!("ing:{done}/{total}"),
-            };
-            recorder.lock().unwrap().push(line);
-        });
-
-        let db = Db::open_in_memory().unwrap();
-        let opts = ScanOptions {
-            jobs: 1,
-            progress: Some(sink),
-            ..Default::default()
-        };
-        let stats = scan_directory_with(&db, dir.path(), &opts).unwrap();
-        assert_eq!(stats.scanned, 5);
-
-        let ev = events.lock().unwrap();
-        // Discovery climbs to the full count.
-        assert!(ev.iter().any(|e| e == "disc:5"), "events: {ev:?}");
-        // Walk reports the total to ingest.
-        assert!(ev.contains(&"walk:5".to_string()), "events: {ev:?}");
-        // Ingest reports each committed file, done strictly 1..=total.
-        let ing: Vec<&String> = ev.iter().filter(|e| e.starts_with("ing:")).collect();
-        assert_eq!(
-            ing,
-            vec!["ing:1/5", "ing:2/5", "ing:3/5", "ing:4/5", "ing:5/5"],
-        );
-    }
-}
-
+mod bounded_probe_tests;
 #[cfg(test)]
-mod ogg_probe_tests {
-    use super::*;
-    use musefs_format::ogg::page_test_support::{
-        build_header_pub, lace_packet_pub, vorbis_body_empty,
-    };
-    use std::io::Write;
-
-    #[test]
-    fn probe_detects_opus_and_seeds_tags() {
-        let head = b"OpusHead\x01\x02\x38\x01\x80\xbb\x00\x00\x00\x00\x00".to_vec();
-        let mut tags = b"OpusTags".to_vec();
-        tags.extend_from_slice(&vorbis_body_empty());
-        let (mut bytes, _) = build_header_pub(0x1234, &[&head, &tags]);
-        let (audio, _) = lace_packet_pub(0x1234, 2, false, 960, &[0u8; 100]);
-        bytes.extend_from_slice(&audio);
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("song.opus");
-        std::fs::File::create(&path)
-            .unwrap()
-            .write_all(&bytes)
-            .unwrap();
-
-        let probed = probe_full(&path, &bytes).expect("opus should probe");
-        assert_eq!(probed.format, Format::Opus);
-        assert_eq!(probed.audio_offset, (bytes.len() - audio.len()) as u64);
-    }
-
-    #[test]
-    fn scan_single_opus_file_ingests_it() {
-        let head = b"OpusHead\x01\x02\x38\x01\x80\xbb\x00\x00\x00\x00\x00".to_vec();
-        let mut tags = b"OpusTags".to_vec();
-        tags.extend_from_slice(&vorbis_body_empty());
-        let (mut bytes, _) = build_header_pub(0x1234, &[&head, &tags]);
-        let (audio, _) = lace_packet_pub(0x1234, 2, false, 960, &[0u8; 100]);
-        bytes.extend_from_slice(&audio);
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("single.opus");
-        std::io::Write::write_all(&mut std::fs::File::create(&path).unwrap(), &bytes).unwrap();
-
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        // Pass the FILE path directly (not the directory).
-        let stats = crate::scan_directory(&db, &path).unwrap();
-        assert_eq!(stats.scanned, 1);
-        assert_eq!(stats.skipped, 0);
-    }
-
-    #[test]
-    fn probe_recognizes_oga_alias() {
-        let head = b"OpusHead\x01\x02\x38\x01\x80\xbb\x00\x00\x00\x00\x00".to_vec();
-        let mut tags = b"OpusTags".to_vec();
-        tags.extend_from_slice(&vorbis_body_empty());
-        let (mut bytes, _) = build_header_pub(0x1234, &[&head, &tags]);
-        let (audio, _) = lace_packet_pub(0x1234, 2, false, 960, &[0u8; 100]);
-        bytes.extend_from_slice(&audio);
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("song.oga");
-        std::fs::File::create(&path)
-            .unwrap()
-            .write_all(&bytes)
-            .unwrap();
-
-        let probed = probe_full(&path, &bytes).expect("oga should probe");
-        assert_eq!(probed.format, Format::Opus);
-    }
-}
-
+mod hardening_tests;
 #[cfg(test)]
-mod wav_probe_tests {
-    use super::*;
-    use std::io::Write;
-
-    fn build_wav() -> Vec<u8> {
-        let mut fmt = Vec::new();
-        fmt.extend_from_slice(&1u16.to_le_bytes());
-        fmt.extend_from_slice(&1u16.to_le_bytes());
-        fmt.extend_from_slice(&44_100u32.to_le_bytes());
-        fmt.extend_from_slice(&88_200u32.to_le_bytes());
-        fmt.extend_from_slice(&2u16.to_le_bytes());
-        fmt.extend_from_slice(&16u16.to_le_bytes());
-
-        let data = vec![0u8; 16];
-        let mut body = Vec::new();
-        for (id, payload) in [(b"fmt ", &fmt), (b"data", &data)] {
-            body.extend_from_slice(id);
-            body.extend_from_slice(&u32::try_from(payload.len()).unwrap().to_le_bytes());
-            body.extend_from_slice(payload);
-        }
-        let mut out = b"RIFF".to_vec();
-        out.extend_from_slice(&u32::try_from(body.len() + 4).unwrap().to_le_bytes());
-        out.extend_from_slice(b"WAVE");
-        out.extend_from_slice(&body);
-        out
-    }
-
-    #[test]
-    fn probe_detects_wav() {
-        let bytes = build_wav();
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("song.wav");
-        std::fs::File::create(&path)
-            .unwrap()
-            .write_all(&bytes)
-            .unwrap();
-
-        let probed = probe_full(&path, &bytes).expect("wav should probe");
-        assert_eq!(probed.format, Format::Wav);
-        assert_eq!(probed.audio_length, 16);
-    }
-
-    #[test]
-    fn scan_single_wav_file_ingests_it() {
-        let bytes = build_wav();
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("single.wav");
-        std::fs::File::create(&path)
-            .unwrap()
-            .write_all(&bytes)
-            .unwrap();
-
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        let stats = crate::scan_directory(&db, &path).unwrap();
-        assert_eq!(stats.scanned, 1);
-        assert_eq!(stats.skipped, 0);
-    }
-}
-
+mod ogg_probe_tests;
 #[cfg(test)]
-mod hardening_tests {
-    use super::*;
-
-    #[test]
-    fn max_art_bytes_is_16_mib_minus_64_kib() {
-        assert_eq!(MAX_ART_BYTES, 16_711_680);
-    }
-
-    #[test]
-    fn scan_caps_match_db_limits() {
-        assert_eq!(
-            i64::try_from(MAX_ART_BYTES).unwrap(),
-            musefs_db::limits::MAX_ART_BYTES
-        );
-        assert_eq!(
-            i64::try_from(MAX_BINARY_TAG_BYTES).unwrap(),
-            musefs_db::limits::MAX_BINARY_TAG_BYTES
-        );
-    }
-
-    #[test]
-    fn is_supported_audio_accepts_known_and_rejects_unknown() {
-        for ok in [
-            "a.flac", "a.mp3", "a.m4a", "a.m4b", "a.ogg", "a.oga", "a.opus", "a.wav",
-        ] {
-            assert!(
-                is_supported_audio(std::path::Path::new(ok)),
-                "{ok} should be supported"
-            );
-        }
-        for bad in ["a.txt", "a.png", "a", "a.flacx"] {
-            assert!(
-                !is_supported_audio(std::path::Path::new(bad)),
-                "{bad} must be rejected"
-            );
-        }
-    }
-
-    #[test]
-    fn collect_audio_skips_unsupported_files() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("keep.flac"), b"x").unwrap();
-        std::fs::write(dir.path().join("skip.txt"), b"x").unwrap();
-        let mut out = Vec::new();
-        collect_audio(dir.path(), &mut out, false).unwrap();
-        assert_eq!(out.len(), 1);
-        assert!(out[0].ends_with("keep.flac"));
-    }
-
-    #[test]
-    fn scan_options_default_does_not_follow_symlinks() {
-        assert!(!ScanOptions::default().follow_symlinks);
-    }
-
-    #[test]
-    fn collect_audio_follows_symlinked_file_when_enabled() {
-        let dir = tempfile::tempdir().unwrap();
-        let real = dir.path().join("real.flac");
-        std::fs::write(&real, b"x").unwrap();
-        let lib = dir.path().join("lib");
-        std::fs::create_dir(&lib).unwrap();
-        std::os::unix::fs::symlink(&real, lib.join("link.flac")).unwrap();
-
-        let mut on = Vec::new();
-        collect_audio(&lib, &mut on, true).unwrap();
-        assert_eq!(
-            on.len(),
-            1,
-            "symlinked file should be collected when following"
-        );
-
-        let mut off = Vec::new();
-        collect_audio(&lib, &mut off, false).unwrap();
-        assert!(
-            off.is_empty(),
-            "symlinked file should be skipped by default"
-        );
-    }
-
-    #[test]
-    fn collect_audio_follows_symlinked_dir_when_enabled() {
-        let dir = tempfile::tempdir().unwrap();
-        let real_dir = dir.path().join("music");
-        std::fs::create_dir(&real_dir).unwrap();
-        std::fs::write(real_dir.join("song.flac"), b"x").unwrap();
-        let root = dir.path().join("root");
-        std::fs::create_dir(&root).unwrap();
-        std::os::unix::fs::symlink(&real_dir, root.join("linkdir")).unwrap();
-
-        let mut on = Vec::new();
-        collect_audio(&root, &mut on, true).unwrap();
-        assert_eq!(
-            on.len(),
-            1,
-            "files under a symlinked dir should be collected"
-        );
-
-        let mut off = Vec::new();
-        collect_audio(&root, &mut off, false).unwrap();
-        assert!(off.is_empty(), "symlinked dir should be skipped by default");
-    }
-
-    #[test]
-    fn collect_audio_terminates_on_symlink_cycle() {
-        let dir = tempfile::tempdir().unwrap();
-        let a = dir.path().join("a");
-        std::fs::create_dir(&a).unwrap();
-        std::fs::write(a.join("song.flac"), b"x").unwrap();
-        std::os::unix::fs::symlink(dir.path(), a.join("loop")).unwrap();
-
-        let mut out = Vec::new();
-        collect_audio(dir.path(), &mut out, true).unwrap();
-        assert_eq!(
-            out.iter().filter(|p| p.ends_with("song.flac")).count(),
-            1,
-            "each real file collected at most once despite the cycle"
-        );
-    }
-
-    #[test]
-    fn collect_audio_skips_broken_symlink_when_following() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("real.flac"), b"x").unwrap();
-        std::os::unix::fs::symlink(dir.path().join("nonexistent"), dir.path().join("dangling"))
-            .unwrap();
-
-        let mut out = Vec::new();
-        let result = collect_audio(dir.path(), &mut out, true);
-        assert!(
-            result.is_ok(),
-            "a dangling symlink must not abort collection"
-        );
-        assert_eq!(out.len(), 1);
-        assert!(out[0].ends_with("real.flac"));
-    }
-
-    #[test]
-    fn collect_audio_does_not_follow_symlinks_by_default() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("real.flac"), b"x").unwrap();
-        let other = dir.path().join("other.flac");
-        std::fs::write(&other, b"x").unwrap();
-        std::os::unix::fs::symlink(&other, dir.path().join("link.flac")).unwrap();
-
-        let mut out = Vec::new();
-        collect_audio(dir.path(), &mut out, false).unwrap();
-        assert_eq!(out.len(), 2);
-    }
-
-    #[test]
-    fn collect_audio_ignores_symlink_to_non_file_target_when_following() {
-        use std::os::unix::ffi::OsStrExt;
-
-        let dir = tempfile::tempdir().unwrap();
-        // A FIFO is neither a regular file nor a directory, and mkfifo works in
-        // restricted sandboxes that deny Unix-socket bind (issue #277).
-        let fifo = dir.path().join("fifo");
-        let c_path = std::ffi::CString::new(fifo.as_os_str().as_bytes()).unwrap();
-        #[expect(unsafe_code, reason = "libc::mkfifo FFI; no std equivalent")]
-        let rc = unsafe { libc::mkfifo(c_path.as_ptr(), 0o644) };
-        assert_eq!(rc, 0, "mkfifo failed: {}", std::io::Error::last_os_error());
-
-        // Name the link with a supported audio extension so the only thing
-        // keeping it out of `out` is the resolved target's is_file() check.
-        std::os::unix::fs::symlink(&fifo, dir.path().join("link.flac")).unwrap();
-
-        let mut out = Vec::new();
-        collect_audio(dir.path(), &mut out, true).unwrap();
-        assert!(
-            out.is_empty(),
-            "a symlink to a non-file, non-dir target must not be collected"
-        );
-    }
-
-    #[test]
-    fn probe_returns_none_for_supported_ext_with_garbage_contents() {
-        let dir = tempfile::tempdir().unwrap();
-        for name in ["bad.flac", "bad.mp3", "bad.m4a", "bad.wav", "bad.opus"] {
-            let path = dir.path().join(name);
-            std::fs::write(&path, b"not a real audio file").unwrap();
-            assert!(
-                probe_full(&path, b"not a real audio file").is_none(),
-                "{name} must skip"
-            );
-        }
-    }
-
-    fn flac_block(bt: u8, body: &[u8], last: bool) -> Vec<u8> {
-        let mut v = vec![(if last { 0x80 } else { 0 }) | (bt & 0x7F)];
-        let n: u32 = u32::try_from(body.len()).unwrap();
-        v.extend_from_slice(&[
-            u8::try_from(n >> 16).unwrap(),
-            u8::try_from(n >> 8).unwrap(),
-            u8::try_from(n).unwrap(),
-        ]);
-        v.extend_from_slice(body);
-        v
-    }
-    fn streaminfo() -> Vec<u8> {
-        let mut si = vec![
-            0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0xC4, 0x42, 0xF0,
-            0x00, 0x00, 0x00, 0x00,
-        ];
-        si.extend_from_slice(&[0u8; 16]);
-        si
-    }
-    fn vorbis_comment(entries: &[&str]) -> Vec<u8> {
-        let mut vc = Vec::new();
-        let vendor = b"x";
-        vc.extend_from_slice(&u32::try_from(vendor.len()).unwrap().to_le_bytes());
-        vc.extend_from_slice(vendor);
-        vc.extend_from_slice(&u32::try_from(entries.len()).unwrap().to_le_bytes());
-        for e in entries {
-            vc.extend_from_slice(&u32::try_from(e.len()).unwrap().to_le_bytes());
-            vc.extend_from_slice(e.as_bytes());
-        }
-        vc
-    }
-    fn picture(width: u32, height: u32, data: &[u8]) -> Vec<u8> {
-        let mut b = Vec::new();
-        b.extend_from_slice(&3u32.to_be_bytes());
-        let mime = "image/png";
-        b.extend_from_slice(&u32::try_from(mime.len()).unwrap().to_be_bytes());
-        b.extend_from_slice(mime.as_bytes());
-        b.extend_from_slice(&0u32.to_be_bytes());
-        b.extend_from_slice(&width.to_be_bytes());
-        b.extend_from_slice(&height.to_be_bytes());
-        b.extend_from_slice(&0u32.to_be_bytes());
-        b.extend_from_slice(&0u32.to_be_bytes());
-        b.extend_from_slice(&u32::try_from(data.len()).unwrap().to_be_bytes());
-        b.extend_from_slice(data);
-        b
-    }
-    fn write_flac(path: &std::path::Path, entries: &[&str], pic: Option<(u32, u32)>) {
-        let mut out = b"fLaC".to_vec();
-        out.extend(flac_block(0, &streaminfo(), false));
-        let last_is_vc = pic.is_none();
-        out.extend(flac_block(4, &vorbis_comment(entries), last_is_vc));
-        if let Some((w, h)) = pic {
-            out.extend(flac_block(6, &picture(w, h, &[0xAB; 64]), true));
-        }
-        out.extend_from_slice(&[0xCD; 128]);
-        std::fs::write(path, &out).unwrap();
-    }
-
-    #[test]
-    fn ingest_assigns_sequential_ordinals_per_key() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("multi.flac");
-        write_flac(&path, &["ARTIST=A1", "ARTIST=A2"], None);
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        crate::scan_directory(&db, &path).unwrap();
-        let track = db.list_tracks().unwrap().into_iter().next().unwrap();
-        let mut artists: Vec<(u64, String)> = db
-            .get_tags(track.id)
-            .unwrap()
-            .into_iter()
-            .filter(|t| t.key.eq_ignore_ascii_case("artist"))
-            .map(|t| (t.ordinal, t.value))
-            .collect();
-        artists.sort();
-        assert_eq!(artists, vec![(0, "A1".to_string()), (1, "A2".to_string())]);
-    }
-
-    #[test]
-    fn ingest_stores_nonzero_art_dimensions() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("art.flac");
-        write_flac(&path, &["ARTIST=A", "TITLE=T"], Some((10, 20)));
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        crate::scan_directory(&db, &path).unwrap();
-        let track = db.list_tracks().unwrap().into_iter().next().unwrap();
-        let ta = db.get_track_art(track.id).unwrap();
-        assert_eq!(ta.len(), 1);
-        let meta = db.get_art_meta(ta[0].art_id).unwrap().unwrap();
-        assert_eq!(meta.width, Some(10));
-        assert_eq!(meta.height, Some(20));
-    }
-
-    #[test]
-    fn ingest_oracle_path_stores_nonzero_art_dimensions() {
-        // Drives the single-file `ingest` (not `ingest_bulk`) so the
-        // `(pic.width != 0).then_some(..)` dimension guards there are pinned.
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("art.flac");
-        write_flac(&path, &["ARTIST=A", "TITLE=T"], Some((10, 20)));
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        crate::scan_directory_full_oracle(&db, &path).unwrap();
-        let track = db.list_tracks().unwrap().into_iter().next().unwrap();
-        let ta = db.get_track_art(track.id).unwrap();
-        assert_eq!(ta.len(), 1);
-        let meta = db.get_art_meta(ta[0].art_id).unwrap().unwrap();
-        assert_eq!(meta.width, Some(10));
-        assert_eq!(meta.height, Some(20));
-    }
-
-    #[test]
-    fn scan_directory_counts_scanned_failed_and_skipped() {
-        let dir = tempfile::tempdir().unwrap();
-        write_flac(
-            &dir.path().join("ok1.flac"),
-            &["ARTIST=A", "TITLE=T1"],
-            None,
-        );
-        write_flac(
-            &dir.path().join("ok2.flac"),
-            &["ARTIST=A", "TITLE=T2"],
-            None,
-        );
-        // Supported extension, unparseable bytes → a scan failure.
-        std::fs::write(dir.path().join("bad.flac"), b"garbage").unwrap();
-        // Unsupported extension → skipped at collection, never probed.
-        std::fs::write(dir.path().join("notes.txt"), b"hello").unwrap();
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        let stats = crate::scan_directory(&db, dir.path()).unwrap();
-        assert_eq!(stats.scanned, 2);
-        assert_eq!(stats.failed, 1);
-        assert_eq!(stats.skipped, 1);
-    }
-
-    #[test]
-    fn skip_tally_summary_orders_by_descending_count() {
-        let mut tally = super::SkipTally::default();
-        for _ in 0..20 {
-            tally.record(std::path::Path::new("art/cover.jpg"));
-        }
-        for _ in 0..10 {
-            tally.record(std::path::Path::new("disc.cue"));
-        }
-        for _ in 0..8 {
-            tally.record(std::path::Path::new("rip.log"));
-        }
-        for _ in 0..4 {
-            tally.record(std::path::Path::new("README"));
-        }
-        assert_eq!(tally.total, 42);
-        assert_eq!(
-            tally.summary().unwrap(),
-            "skipped 42: jpg=20, cue=10, log=8, <none>=4"
-        );
-    }
-
-    #[test]
-    fn skip_tally_lowercases_extension_and_buckets_extensionless() {
-        let mut tally = super::SkipTally::default();
-        tally.record(std::path::Path::new("a.JPG"));
-        tally.record(std::path::Path::new("b.jpg"));
-        tally.record(std::path::Path::new("noext"));
-        assert_eq!(tally.summary().unwrap(), "skipped 3: jpg=2, <none>=1");
-    }
-
-    #[test]
-    fn skip_tally_ties_break_by_extension_name() {
-        let mut tally = super::SkipTally::default();
-        tally.record(std::path::Path::new("a.nfo"));
-        tally.record(std::path::Path::new("b.cue"));
-        assert_eq!(tally.summary().unwrap(), "skipped 2: cue=1, nfo=1");
-    }
-
-    #[test]
-    fn skip_tally_empty_has_no_summary() {
-        assert!(super::SkipTally::default().summary().is_none());
-    }
-
-    #[test]
-    fn revalidate_buckets_unchanged_and_prunes_missing() {
-        let dir = tempfile::tempdir().unwrap();
-        let keep = dir.path().join("keep.flac");
-        write_flac(&keep, &["ARTIST=A", "TITLE=T"], None);
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        crate::scan_directory(&db, dir.path()).unwrap();
-
-        let s1 = crate::revalidate(&db, dir.path()).unwrap();
-        assert_eq!(s1.unchanged, 1);
-        assert_eq!(s1.updated, 0);
-        assert_eq!(s1.pruned, 0);
-
-        std::fs::remove_file(&keep).unwrap();
-        let s2 = crate::revalidate(&db, dir.path()).unwrap();
-        assert_eq!(s2.pruned, 1);
-        assert!(db.list_tracks().unwrap().is_empty());
-    }
-
-    #[test]
-    fn revalidate_does_not_prune_on_non_notfound_error() {
-        let dir = tempfile::tempdir().unwrap();
-        let file = dir.path().join("real.flac");
-        write_flac(&file, &["ARTIST=A", "TITLE=T"], None);
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        crate::scan_directory(&db, dir.path()).unwrap();
-
-        use musefs_db::{Format, NewTrack};
-        let track = db.list_tracks().unwrap().into_iter().next().unwrap();
-        db.delete_track(track.id).unwrap();
-        let canon = std::fs::canonicalize(dir.path()).unwrap();
-        let ghost = canon.join("real.flac").join("ghost.flac");
-        db.upsert_track(&NewTrack {
-            backing_path: ghost.to_string_lossy().into_owned(),
-            format: Format::Flac,
-            audio_offset: 0,
-            audio_length: 0,
-            backing_size: 0,
-            backing_mtime_ns: 0,
-            backing_ctime_ns: 0,
-        })
-        .unwrap();
-
-        let stats = crate::revalidate(&db, dir.path()).unwrap();
-        assert_eq!(stats.pruned, 0, "ENOTDIR is not NotFound → must not prune");
-        assert!(
-            db.list_tracks()
-                .unwrap()
-                .iter()
-                .any(|t| t.backing_path == ghost.to_string_lossy()),
-            "ghost track must still exist"
-        );
-    }
-
-    #[test]
-    fn scan_ingests_binary_tags_and_promotes() {
-        use id3::frame::{Content, Popularimeter, Unknown};
-        use id3::{Encoder, Frame, Tag, TagLike, Version};
-
-        let dir = tempfile::tempdir().unwrap();
-
-        // Build an MP3 with a PRIV (opaque) + POPM (promoted) tag.
-        let mut tag = Tag::new();
-        tag.add_frame(Popularimeter {
-            user: "u".into(),
-            rating: 128,
-            counter: 3,
-        });
-        tag.add_frame(Frame::with_content(
-            "PRIV",
-            Content::Unknown(Unknown {
-                data: vec![1, 1, 2, 3, 5],
-                version: Version::Id3v24,
-            }),
-        ));
-        let mut bytes = Vec::new();
-        Encoder::new()
-            .version(Version::Id3v24)
-            .encode(&tag, &mut bytes)
-            .unwrap();
-        // A real MP3 frame header is enough for locate_audio_bounded to find audio.
-        bytes.extend_from_slice(&[0xFF, 0xFB, 0x90, 0x00, 0x00, 0x00, 0x00, 0x00]);
-        std::fs::write(dir.path().join("a.mp3"), &bytes).unwrap();
-
-        let db = musefs_db::Db::open_in_memory().unwrap();
-        crate::scan::scan_directory(&db, dir.path()).unwrap();
-        let track = db.list_tracks().unwrap().into_iter().next().unwrap();
-        let tid = track.id;
-
-        // Opaque PRIV survives as a binary row.
-        let bin = db.get_binary_tags(tid).unwrap();
-        assert!(
-            bin.iter().any(|r| r.key == "PRIV" && r.byte_len == 5),
-            "PRIV not ingested as binary row; got: {bin:?}"
-        );
-
-        // POPM promoted into editable text tags.
-        let texts = db.get_tags(tid).unwrap();
-        assert!(
-            texts.iter().any(|t| t.key == "rating" && t.value == "128"),
-            "rating not promoted; got: {texts:?}"
-        );
-        assert!(
-            texts.iter().any(|t| t.key == "playcount" && t.value == "3"),
-            "playcount not promoted; got: {texts:?}"
-        );
-    }
-
-    /// Probed carrying a valid, an empty, and an oversize binary tag. Only the
-    /// valid one is stored: the filter drops empty (`EmptySegment` would fail
-    /// layout validation) and oversize (`> MAX_BINARY_TAG_BYTES`) payloads, with
-    /// gap-free ordinals.
-    fn probed_with_mixed_binary_tags() -> Probed {
-        Probed {
-            format: musefs_db::Format::Mp3,
-            audio_offset: 0,
-            audio_length: 0,
-            tags: Vec::new(),
-            pictures: Vec::new(),
-            binary_tags: vec![
-                EmbeddedBinaryTag {
-                    key: "PRIV".into(),
-                    payload: vec![1, 2, 3],
-                },
-                EmbeddedBinaryTag {
-                    key: "GEOB".into(),
-                    payload: Vec::new(),
-                },
-                EmbeddedBinaryTag {
-                    key: "SYLT".into(),
-                    payload: vec![0u8; MAX_BINARY_TAG_BYTES + 1],
-                },
-            ],
-            structural_blocks: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn ingest_filters_empty_and_oversize_binary_tags() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("a.mp3");
-        std::fs::write(&path, b"x").unwrap();
-        let meta = std::fs::metadata(&path).unwrap();
-        let db = Db::open_in_memory().unwrap();
-
-        ingest(
-            &db,
-            &path.to_string_lossy(),
-            &meta,
-            probed_with_mixed_binary_tags(),
-        )
-        .unwrap();
-
-        let tid = db.list_tracks().unwrap()[0].id;
-        let rows = db.get_binary_tags(tid).unwrap();
-        assert_eq!(
-            rows.len(),
-            1,
-            "only the valid binary tag survives: {rows:?}"
-        );
-        assert_eq!(rows[0].key, "PRIV");
-        assert_eq!(rows[0].byte_len, 3);
-    }
-
-    #[test]
-    fn ingest_bulk_filters_empty_and_oversize_binary_tags() {
-        let db = Db::open_in_memory().unwrap();
-        {
-            let mut bw = db.bulk_writer().unwrap();
-            ingest_bulk(
-                &mut bw,
-                "/a.mp3",
-                BackingStamp {
-                    size: 1,
-                    mtime_ns: 0,
-                    ctime_ns: 0,
-                },
-                probed_with_mixed_binary_tags(),
-            )
-            .unwrap();
-            bw.commit().unwrap();
-        }
-        let tid = db.list_tracks().unwrap()[0].id;
-        let rows = db.get_binary_tags(tid).unwrap();
-        assert_eq!(
-            rows.len(),
-            1,
-            "only the valid binary tag survives: {rows:?}"
-        );
-        assert_eq!(rows[0].key, "PRIV");
-        assert_eq!(rows[0].byte_len, 3);
-    }
-
-    #[test]
-    fn accept_pictures_keeps_at_cap_and_drops_over_cap() {
-        let mk = |len: usize| EmbeddedPicture {
-            mime: "image/jpeg".to_string(),
-            picture_type: musefs_format::PictureType::new(3).unwrap(),
-            description: String::new(),
-            width: 0,
-            height: 0,
-            data: vec![0u8; len],
-        };
-        // A picture exactly at the cap is kept; one byte over is dropped. The
-        // boundary pins `>` against `>=` (an at-cap drop would be silent loss).
-        let kept = accept_pictures("/x.flac", vec![mk(MAX_ART_BYTES), mk(MAX_ART_BYTES + 1)]);
-        assert_eq!(kept.len(), 1, "exactly the at-cap picture survives");
-        assert_eq!(kept[0].data.len(), MAX_ART_BYTES);
-    }
-
-    #[test]
-    fn accept_binary_tags_keeps_at_cap_and_drops_over_cap() {
-        let mk = |len: usize| EmbeddedBinaryTag {
-            key: "PRIV".to_string(),
-            payload: vec![0u8; len],
-        };
-        let kept = accept_binary_tags(
-            "/x.mp3",
-            vec![mk(MAX_BINARY_TAG_BYTES), mk(MAX_BINARY_TAG_BYTES + 1)],
-        );
-        assert_eq!(kept.len(), 1, "exactly the at-cap binary tag survives");
-        assert_eq!(kept[0].payload.len(), MAX_BINARY_TAG_BYTES);
-    }
-
-    fn probed_with_text_tags(tags: &[(&str, &str)]) -> Probed {
-        Probed {
-            format: musefs_db::Format::Mp3,
-            audio_offset: 0,
-            audio_length: 0,
-            tags: tags
-                .iter()
-                .map(|(k, v)| ((*k).to_string(), (*v).to_string()))
-                .collect(),
-            pictures: Vec::new(),
-            binary_tags: Vec::new(),
-            structural_blocks: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn ingest_skips_empty_and_control_char_keys() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("a.mp3");
-        std::fs::write(&path, b"x").unwrap();
-        let meta = std::fs::metadata(&path).unwrap();
-        let db = Db::open_in_memory().unwrap();
-
-        ingest(
-            &db,
-            &path.to_string_lossy(),
-            &meta,
-            probed_with_text_tags(&[
-                ("artist", "Alice"),
-                ("", "dropped"),        // empty key
-                ("a\u{7}b", "dropped"), // control char
-                ("a\u{0}b", "dropped"), // embedded NUL — DB CHECK can't see it, the floor can
-                ("a=b", "kept"),        // '=' is NOT a floor violation
-            ]),
-        )
-        .unwrap();
-
-        let tid = db.list_tracks().unwrap()[0].id;
-        let keys: Vec<String> = db
-            .get_tags(tid)
-            .unwrap()
-            .into_iter()
-            .map(|t| t.key)
-            .collect();
-        // get_tags is ORDER BY key, ordinal: '=' (0x3D) sorts before 'a' (0x61).
-        assert_eq!(keys, vec!["a=b".to_string(), "artist".to_string()]);
-    }
-
-    #[test]
-    fn ingest_bulk_skips_empty_and_control_char_keys() {
-        let db = Db::open_in_memory().unwrap();
-        {
-            let mut bw = db.bulk_writer().unwrap();
-            ingest_bulk(
-                &mut bw,
-                "/a.mp3",
-                BackingStamp {
-                    size: 1,
-                    mtime_ns: 0,
-                    ctime_ns: 0,
-                },
-                probed_with_text_tags(&[
-                    ("artist", "Alice"),
-                    ("", "dropped"),
-                    ("a\u{7}b", "dropped"),
-                    ("a\u{0}b", "dropped"), // embedded NUL — floor drops it
-                    ("a=b", "kept"),
-                ]),
-            )
-            .unwrap();
-            bw.commit().unwrap();
-        }
-        let tid = db.list_tracks().unwrap()[0].id;
-        let keys: Vec<String> = db
-            .get_tags(tid)
-            .unwrap()
-            .into_iter()
-            .map(|t| t.key)
-            .collect();
-        assert_eq!(keys, vec!["a=b".to_string(), "artist".to_string()]);
-    }
-
-    /// Probed with two structural blocks of the SAME kind, to make the per-kind
-    /// ordinal increment (`*ord += 1`) observable. A real FLAC carries only one
-    /// STREAMINFO/SEEKTABLE, so a duplicate kind is the only input under which the
-    /// second block's ordinal differs from the first; without it the increment's
-    /// mutants survive.
-    fn probed_with_duplicate_structural_kind() -> Probed {
-        Probed {
-            format: musefs_db::Format::Flac,
-            audio_offset: 0,
-            audio_length: 0,
-            tags: Vec::new(),
-            pictures: Vec::new(),
-            binary_tags: Vec::new(),
-            structural_blocks: vec![
-                ("SEEKTABLE".to_string(), vec![0xA1]),
-                ("SEEKTABLE".to_string(), vec![0xB2]),
-            ],
-        }
-    }
-
-    #[test]
-    fn ingest_assigns_sequential_structural_ordinals_per_kind() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("a.flac");
-        std::fs::write(&path, b"x").unwrap();
-        let meta = std::fs::metadata(&path).unwrap();
-        let db = Db::open_in_memory().unwrap();
-
-        ingest(
-            &db,
-            &path.to_string_lossy(),
-            &meta,
-            probed_with_duplicate_structural_kind(),
-        )
-        .unwrap();
-
-        let tid = db.list_tracks().unwrap()[0].id;
-        let got = db.get_structural_blocks(tid).unwrap();
-        // Rows come back ORDER BY kind, ordinal: the two same-kind blocks must hold
-        // ordinals 0 then 1 (the `-=`/`*=` mutants collapse or invert this).
-        assert_eq!(got.len(), 2);
-        assert_eq!(got[0].ordinal, 0);
-        assert_eq!(got[0].body, vec![0xA1]);
-        assert_eq!(got[1].ordinal, 1);
-        assert_eq!(got[1].body, vec![0xB2]);
-    }
-
-    /// Probed with two tags of the SAME key, to make the per-key ordinal
-    /// increment (`*ord += 1` in the tag loop) observable. The production
-    /// `ingest_bulk` path is exercised with a multi-value tag elsewhere, but the
-    /// oracle-only `ingest` is not, so without this its tag-ordinal mutants
-    /// survive. Distinct values under one key: a collapsed ordinal (the `-=`/`*=`
-    /// mutants) either underflows or duplicates the `(track_id, key, ordinal)`
-    /// primary key — both observable.
-    fn probed_with_duplicate_tag_key() -> Probed {
-        Probed {
-            format: musefs_db::Format::Flac,
-            audio_offset: 0,
-            audio_length: 0,
-            tags: vec![
-                ("ARTIST".to_string(), "A".to_string()),
-                ("ARTIST".to_string(), "B".to_string()),
-            ],
-            pictures: Vec::new(),
-            binary_tags: Vec::new(),
-            structural_blocks: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn ingest_assigns_sequential_tag_ordinals_per_key() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("a.flac");
-        std::fs::write(&path, b"x").unwrap();
-        let meta = std::fs::metadata(&path).unwrap();
-        let db = Db::open_in_memory().unwrap();
-
-        ingest(
-            &db,
-            &path.to_string_lossy(),
-            &meta,
-            probed_with_duplicate_tag_key(),
-        )
-        .unwrap();
-
-        let tid = db.list_tracks().unwrap()[0].id;
-        let got = db.get_tags(tid).unwrap();
-        // get_tags is ORDER BY key, ordinal: the two same-key tags must hold
-        // ordinals 0 then 1 (the `-=`/`*=` mutants collapse or invert this).
-        assert_eq!(got.len(), 2);
-        assert_eq!(got[0].ordinal, 0);
-        assert_eq!(got[0].value, "A");
-        assert_eq!(got[1].ordinal, 1);
-        assert_eq!(got[1].value, "B");
-    }
-
-    #[test]
-    fn ingest_bulk_assigns_sequential_structural_ordinals_per_kind() {
-        let db = Db::open_in_memory().unwrap();
-        {
-            let mut bw = db.bulk_writer().unwrap();
-            ingest_bulk(
-                &mut bw,
-                "/a.flac",
-                BackingStamp {
-                    size: 1,
-                    mtime_ns: 0,
-                    ctime_ns: 0,
-                },
-                probed_with_duplicate_structural_kind(),
-            )
-            .unwrap();
-            bw.commit().unwrap();
-        }
-        let tid = db.list_tracks().unwrap()[0].id;
-        let got = db.get_structural_blocks(tid).unwrap();
-        assert_eq!(got.len(), 2);
-        assert_eq!(got[0].ordinal, 0);
-        assert_eq!(got[0].body, vec![0xA1]);
-        assert_eq!(got[1].ordinal, 1);
-        assert_eq!(got[1].body, vec![0xB2]);
-    }
-}
-
+mod scan_unit_tests;
 #[cfg(test)]
-mod bounded_probe_tests {
-    use super::*;
-    use musefs_db::Db;
-
-    /// Minimal FLAC: marker + a single last STREAMINFO (34-byte body) + audio.
-    /// FLAC has no frame-sync check at the audio offset, so any payload works.
-    fn flac_fixture() -> Vec<u8> {
-        let mut bytes = b"fLaC".to_vec();
-        bytes.push(0x80); // last-block flag set, type 0 (STREAMINFO)
-        bytes.extend_from_slice(&[0, 0, 34]); // 24-bit length = 34
-        bytes.extend(std::iter::repeat_n(0u8, 34));
-        bytes.extend_from_slice(b"AUDIOPAYLOAD");
-        bytes
-    }
-
-    #[test]
-    fn scan_counts_unreadable_file_as_failed_and_continues() {
-        let dir = tempfile::tempdir().unwrap();
-        // One good FLAC + one zero-byte ".flac" that cannot parse.
-        let good = dir.path().join("good.flac");
-        let mut bytes = b"fLaC".to_vec();
-        bytes.push(0x80);
-        bytes.extend_from_slice(&[0, 0, 34]);
-        bytes.extend(std::iter::repeat_n(0u8, 34));
-        bytes.extend_from_slice(b"AUDIO");
-        std::fs::write(&good, &bytes).unwrap();
-        std::fs::write(dir.path().join("bad.flac"), b"").unwrap();
-
-        let db = Db::open_in_memory().unwrap();
-        let stats = scan_directory(&db, dir.path()).unwrap();
-        assert_eq!(stats.scanned, 1);
-        assert_eq!(stats.skipped + stats.failed, 1);
-    }
-
-    #[test]
-    fn scan_directory_bounded_matches_full_for_flac() {
-        // A FLAC fixture written to a temp dir, scanned with the (default) bounded
-        // path, yields a track with the same audio bounds as a full-file probe.
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("a.flac");
-        let bytes = flac_fixture();
-        std::fs::write(&path, &bytes).unwrap();
-
-        let full = probe_full(&path, &bytes).expect("full probe");
-
-        let db = Db::open_in_memory().unwrap();
-        let stats = scan_directory(&db, dir.path()).unwrap();
-        assert_eq!(stats.scanned, 1);
-        let track = db
-            .get_track_by_path(&std::fs::canonicalize(&path).unwrap().to_string_lossy())
-            .unwrap()
-            .unwrap();
-        assert_eq!(track.bounds.audio_offset(), full.audio_offset);
-        assert_eq!(track.bounds.audio_length(), full.audio_length);
-    }
-
-    #[test]
-    fn revalidate_skips_unchanged_and_reprobes_changed() {
-        let dir = tempfile::tempdir().unwrap();
-        let p = dir.path().join("x.flac");
-        let mk = |audio: &[u8]| {
-            let mut b = b"fLaC".to_vec();
-            b.push(0x80);
-            b.extend_from_slice(&[0, 0, 34]);
-            b.extend(std::iter::repeat_n(0u8, 34));
-            b.extend_from_slice(audio);
-            b
-        };
-        std::fs::write(&p, mk(b"AUDIO")).unwrap();
-        let db = Db::open_in_memory().unwrap();
-        scan_directory(&db, dir.path()).unwrap();
-
-        // Unchanged → all unchanged.
-        let s1 = revalidate_with(&db, dir.path(), &ScanOptions::default()).unwrap();
-        assert_eq!(s1.unchanged, 1);
-        assert_eq!(s1.updated, 0);
-
-        // Rewrite with a different size → detected as changed and re-probed.
-        std::fs::write(&p, mk(b"DIFFERENT-AUDIO")).unwrap();
-        let s2 = revalidate_with(&db, dir.path(), &ScanOptions::default()).unwrap();
-        assert_eq!(s2.updated, 1);
-        assert_eq!(s2.unchanged, 0);
-        // The track row now reflects the new (longer) audio length.
-        let track = db
-            .get_track_by_path(&std::fs::canonicalize(&p).unwrap().to_string_lossy())
-            .unwrap()
-            .unwrap();
-        assert_eq!(
-            usize_from(track.bounds.audio_length()),
-            b"DIFFERENT-AUDIO".len()
-        );
-    }
-
-    #[test]
-    fn revalidate_accepts_a_single_file_target() {
-        // The CLI advertises file targets for every scan, including --revalidate,
-        // so revalidate_with must handle a bare file root (not just a directory).
-        let dir = tempfile::tempdir().unwrap();
-        let p = dir.path().join("x.flac");
-        let mut bytes = b"fLaC".to_vec();
-        bytes.push(0x80);
-        bytes.extend_from_slice(&[0, 0, 34]);
-        bytes.extend(std::iter::repeat_n(0u8, 34));
-        bytes.extend_from_slice(b"AUDIO");
-        std::fs::write(&p, &bytes).unwrap();
-        let db = Db::open_in_memory().unwrap();
-        scan_directory(&db, dir.path()).unwrap();
-
-        // Revalidate the file path directly: must not error on read_dir and the
-        // unchanged file is bucketed as unchanged (not pruned).
-        let stats = revalidate_with(&db, &p, &ScanOptions::default()).unwrap();
-        assert_eq!(stats.unchanged, 1);
-        assert_eq!(stats.pruned, 0);
-        assert_eq!(db.list_tracks().unwrap().len(), 1);
-    }
-
-    #[test]
-    fn jobs1_and_jobs_n_produce_equivalent_state() {
-        let dir = tempfile::tempdir().unwrap();
-        // A handful of distinct FLACs.
-        for i in 0..12 {
-            let mut bytes = b"fLaC".to_vec();
-            bytes.push(0x80);
-            bytes.extend_from_slice(&[0, 0, 34]);
-            bytes.extend(std::iter::repeat_n(0u8, 34));
-            bytes.extend_from_slice(format!("AUDIO-{i}").as_bytes());
-            std::fs::write(dir.path().join(format!("t{i}.flac")), &bytes).unwrap();
-        }
-        let norm = |jobs: usize| {
-            let db = Db::open_in_memory().unwrap();
-            scan_directory_with(
-                &db,
-                dir.path(),
-                &ScanOptions {
-                    jobs,
-                    ..Default::default()
-                },
-            )
-            .unwrap();
-            let mut rows: Vec<(String, u64, u64)> = db
-                .list_tracks()
-                .unwrap()
-                .into_iter()
-                .map(|t| {
-                    (
-                        t.backing_path,
-                        t.bounds.audio_offset(),
-                        t.bounds.audio_length(),
-                    )
-                })
-                .collect();
-            rows.sort();
-            rows
-        };
-        assert_eq!(norm(1), norm(4));
-        assert_eq!(norm(1).len(), 12);
-    }
-
-    #[test]
-    fn oversize_unparseable_file_is_skipped_not_read_whole() {
-        // A file far larger than the probe ceiling, with a valid FLAC marker but
-        // a metadata block that never terminates, must be skipped rather than
-        // allocated whole into RAM (the misnamed-multi-GB-file OOM guard).
-        use std::io::Write;
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("huge.flac");
-        let mut f = std::fs::File::create(&path).unwrap();
-        // Marker + a non-last VORBIS_COMMENT block claiming the max 24-bit
-        // length, so the bounded reader keeps asking for more.
-        f.write_all(b"fLaC").unwrap();
-        f.write_all(&[0x04, 0xFF, 0xFF, 0xFF]).unwrap();
-        let len = MAX_PROBE_BYTES + 4096;
-        f.set_len(len).unwrap();
-        drop(f);
-
-        assert!(matches!(
-            probe_file(&path, WINDOW).unwrap(),
-            ProbeOutcome::Unparseable
-        ));
-    }
-
-    #[test]
-    fn oversize_wav_is_served_via_data_header() {
-        // A valid WAV whose `data` payload exceeds the probe ceiling (any
-        // recording more than a few minutes long) must still be ingested: the
-        // `data` chunk header sits at the front, so the declared audio bounds
-        // are known without reading the payload. Skipping it would drop every
-        // sufficiently long WAV in the library.
-        use std::io::Write;
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("long.wav");
-
-        let data_len: u64 = MAX_PROBE_BYTES + (16 << 20); // 80 MiB payload
-        let mut fmt = Vec::new();
-        fmt.extend_from_slice(&1u16.to_le_bytes());
-        fmt.extend_from_slice(&1u16.to_le_bytes());
-        fmt.extend_from_slice(&44_100u32.to_le_bytes());
-        fmt.extend_from_slice(&88_200u32.to_le_bytes());
-        fmt.extend_from_slice(&2u16.to_le_bytes());
-        fmt.extend_from_slice(&16u16.to_le_bytes());
-
-        let mut front = b"RIFF".to_vec();
-        // form: WAVE(4) + fmt chunk(24) + data header(8) + data payload
-        let riff_size = 36u32 + u32::try_from(data_len).unwrap();
-        front.extend_from_slice(&riff_size.to_le_bytes());
-        front.extend_from_slice(b"WAVE");
-        front.extend_from_slice(b"fmt ");
-        front.extend_from_slice(&u32::try_from(fmt.len()).unwrap().to_le_bytes());
-        front.extend_from_slice(&fmt);
-        front.extend_from_slice(b"data");
-        front.extend_from_slice(&u32::try_from(data_len).unwrap().to_le_bytes());
-        let audio_offset = front.len() as u64;
-        let file_len = audio_offset + data_len;
-
-        let mut f = std::fs::File::create(&path).unwrap();
-        f.write_all(&front).unwrap();
-        f.set_len(file_len).unwrap();
-        drop(f);
-
-        let probed = match probe_file(&path, WINDOW).unwrap() {
-            ProbeOutcome::Probed(p, _) => p,
-            other => panic!("expected Probed, got {other:?}"),
-        };
-        assert_eq!(probed.format, Format::Wav);
-        assert_eq!(probed.audio_offset, audio_offset);
-        assert_eq!(probed.audio_length, data_len);
-    }
-
-    #[test]
-    fn probe_file_reports_raced_on_mid_probe_mutation() {
-        use std::io::Write;
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("a.wav");
-
-        // Minimal valid WAV the probe accepts (fmt + tiny data).
-        let mut fmt = Vec::new();
-        for v in [1u16, 1, 0, 0, 0, 16] {
-            fmt.extend_from_slice(&v.to_le_bytes());
-        }
-        let mut front = b"RIFF".to_vec();
-        // form: WAVE(4) + fmt chunk(8+len) + data header(8) + data payload(64)
-        let riff_size = 4 + 8 + u32::try_from(fmt.len()).unwrap() + 8 + 64;
-        front.extend_from_slice(&riff_size.to_le_bytes());
-        front.extend_from_slice(b"WAVE");
-        front.extend_from_slice(b"fmt ");
-        front.extend_from_slice(&u32::try_from(fmt.len()).unwrap().to_le_bytes());
-        front.extend_from_slice(&fmt);
-        front.extend_from_slice(b"data");
-        front.extend_from_slice(&64u32.to_le_bytes());
-        let mut f = std::fs::File::create(&path).unwrap();
-        f.write_all(&front).unwrap();
-        f.set_len(front.len() as u64 + 64).unwrap();
-        drop(f);
-
-        let pc = path.clone();
-        set_after_s1_hook(move || {
-            let mut g = std::fs::OpenOptions::new().append(true).open(&pc).unwrap();
-            g.write_all(&[0u8; 4096]).unwrap(); // size moves -> S2 != S1
-        });
-        let out = probe_file(&path, WINDOW);
-        clear_after_s1_hook();
-        assert!(matches!(out, Ok(ProbeOutcome::Raced)), "got {out:?}");
-    }
-}
+mod wav_probe_tests;
