+++ musefs-core/src/scan/wav_probe_tests.rs
@@ -1,0 +1,56 @@
+use super::*;
+use std::io::Write;
+
+fn build_wav() -> Vec<u8> {
+    let mut fmt = Vec::new();
+    fmt.extend_from_slice(&1u16.to_le_bytes());
+    fmt.extend_from_slice(&1u16.to_le_bytes());
+    fmt.extend_from_slice(&44_100u32.to_le_bytes());
+    fmt.extend_from_slice(&88_200u32.to_le_bytes());
+    fmt.extend_from_slice(&2u16.to_le_bytes());
+    fmt.extend_from_slice(&16u16.to_le_bytes());
+
+    let data = vec![0u8; 16];
+    let mut body = Vec::new();
+    for (id, payload) in [(b"fmt ", &fmt), (b"data", &data)] {
+        body.extend_from_slice(id);
+        body.extend_from_slice(&u32::try_from(payload.len()).unwrap().to_le_bytes());
+        body.extend_from_slice(payload);
+    }
+    let mut out = b"RIFF".to_vec();
+    out.extend_from_slice(&u32::try_from(body.len() + 4).unwrap().to_le_bytes());
+    out.extend_from_slice(b"WAVE");
+    out.extend_from_slice(&body);
+    out
+}
+
+#[test]
+fn probe_detects_wav() {
+    let bytes = build_wav();
+    let dir = tempfile::tempdir().unwrap();
+    let path = dir.path().join("song.wav");
+    std::fs::File::create(&path)
+        .unwrap()
+        .write_all(&bytes)
+        .unwrap();
+
+    let probed = probe_full(&path, &bytes).expect("wav should probe");
+    assert_eq!(probed.format, Format::Wav);
+    assert_eq!(probed.audio_length, 16);
+}
+
+#[test]
+fn scan_single_wav_file_ingests_it() {
+    let bytes = build_wav();
+    let dir = tempfile::tempdir().unwrap();
+    let path = dir.path().join("single.wav");
+    std::fs::File::create(&path)
+        .unwrap()
+        .write_all(&bytes)
+        .unwrap();
+
+    let db = musefs_db::Db::open_in_memory().unwrap();
+    let stats = crate::scan_directory(&db, &path).unwrap();
+    assert_eq!(stats.scanned, 1);
+    assert_eq!(stats.skipped, 0);
+}
